# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 晟風工程 ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /work-orders:
    get:
      operationId: listWorkOrders
      tags: [workOrders]
      summary: 列出所有派工單
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 派工單列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
    post:
      operationId: createWorkOrder
      tags: [workOrders]
      summary: 新增派工單
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderInput"
      responses:
        "201":
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /work-orders/{id}:
    get:
      operationId: getWorkOrder
      tags: [workOrders]
      summary: 取得派工單詳細資料
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 派工單資料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "404":
          description: 找不到派工單
    patch:
      operationId: updateWorkOrder
      tags: [workOrders]
      summary: 更新派工單
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderUpdate"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "404":
          description: 找不到派工單
    delete:
      operationId: deleteWorkOrder
      tags: [workOrders]
      summary: 刪除派工單
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 刪除成功
  /work-orders/{workOrderId}/progress:
    get:
      operationId: listProgress
      tags: [progress]
      summary: 取得派工單的工程進度
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 工程進度列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Progress"
    post:
      operationId: createProgress
      tags: [progress]
      summary: 新增工程進度紀錄
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressInput"
      responses:
        "201":
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
...
components:
  schemas:
...
    WorkOrder:
      type: object
      required: [id, customerId, title, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        quoteId: { type: ["integer", "null"] }
        title: { type: string }
        description: { type: ["string", "null"] }
        assignedTo: { type: ["string", "null"] }
        scheduledDate: { type: ["string", "null"] }
        completedDate: { type: ["string", "null"] }
        status: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    WorkOrderInput:
      type: object
      required: [customerId, title, status]
      properties:
        customerId: { type: integer }
        quoteId: { type: integer }
        title: { type: string, minLength: 1 }
        description: { type: string }
        assignedTo: { type: string }
        scheduledDate: { type: string }
        completedDate: { type: string }
        status: { type: string }
        notes: { type: string }
    WorkOrderUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        assignedTo: { type: string }
        scheduledDate: { type: string }
        completedDate: { type: string }
        status: { type: string }
        notes: { type: string }
    Progress:
      type: object
      required: [id, workOrderId, description, createdAt]
      properties:
        id: { type: integer }
        workOrderId: { type: integer }
        description: { type: string }
        recordedBy: { type: ["string", "null"] }
        createdAt: { type: string }
    ProgressInput:
      type: object
      required: [description]
      properties:
        description: { type: string, minLength: 1 }
        recordedBy: { type: string }
...