# lib/api-zod/src/generated/types/workOrder.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 晟風工程 ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkOrder {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  quoteId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  scheduledDate?: string | null;
  /** @nullable */
  completedDate?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}