# lib/api-zod/src/generated/types/workOrderUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 晟風工程 ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkOrderUpdate {
  title?: string;
  description?: string;
  assignedTo?: string;
  scheduledDate?: string;
  completedDate?: string;
  status?: string;
  notes?: string;
}